Don't use compile-time `long double` arithmetic in src/gentables

+++ src/gentables/fluid_concave.cpp
@@ -14,7 +14,7 @@ struct ConcaveFunctor
             ? 0
             : ((i == FLUID_VEL_CB_SIZE - 1)
                 ? 1
-                : ((-200.0L * 2 / FLUID_PEAK_ATTENUATION) * gcem::log(((FLUID_VEL_CB_SIZE - 1) - i) / (FLUID_VEL_CB_SIZE - 1.0L)) / GCEM_LOG_10)
+                : ((-200.0 * 2 / FLUID_PEAK_ATTENUATION) * gcem::log(((FLUID_VEL_CB_SIZE - 1) - i) / (FLUID_VEL_CB_SIZE - 1.0)) / static_cast<double>(GCEM_LOG_10))
                 ));
     }
 };
